requirements.txt
lxml==4.9.2
beautifulsoup4==4.12.2
requests==2.28.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pytest==7.3.1
fastapi==0.95.1 
uvicorn==0.22.0